pyproject.toml
[project]
name = "enterprise-study-env"
version = "0.1.0"
description = "AuDHD Enterprise Study Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=1.2.20",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "langgraph>=0.1.18",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "py2app>=0.28.10",
    "pydantic>=2.12.5",
    "pyinstaller>=6.19.0",
    "pyobjc>=12.1",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "rumps>=0.4.0",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "typer>=0.24.1",
]

[project.scripts]
audhd-pipeline = "zennode.cli.main:app"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B", "A", "C4", "SIM", "DTZ", "PT", "RET"]
ignore = []

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zennode"]
